refactor: fileSystem out of cli
